tests/requirements-test.txt
# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # For parallel test execution

# API Testing
requests>=2.31.0
httpx>=0.24.0

# Docker Testing
docker>=6.1.0
testcontainers>=3.7.0

# GCP SDKs
google-cloud-aiplatform>=1.35.0
google-cloud-logging>=3.5.0
google-cloud-monitoring>=2.15.0
google-cloud-run>=0.10.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0

# VertexAI & Gemini
google-generativeai>=0.3.0
vertexai>=1.35.0

# Performance Testing
locust>=2.15.0  # For load testing

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
dnspython>=2.4.0  # For DNS testing
pyOpenSSL>=23.2.0  # For SSL testing

# Data Analysis (for metrics)
numpy>=1.24.0
pandas>=2.0.0  # Optional, for advanced analysis
